fixes in namespace + plugin header + in old legacy code (not addressing fixes in memory management in legacy code though, should be done in independent PR).
